Refactored all the codes from App.js to Home.js and wrapped the codes in navigationContainer
